[feature/#211] board 임시 수정

+++ resource-server/src/main/java/com/inhabas/api/domain/board/repository/NormalBoardRepository.java
@@ -3,5 +3,5 @@ package com.inhabas.api.domain.board.repository;
 import com.inhabas.api.domain.board.domain.NormalBoard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NormalBoardRepository extends JpaRepository<NormalBoard, Integer>, NormalBoardRepositoryCustom {
+public interface NormalBoardRepository extends JpaRepository<NormalBoard, Long>, NormalBoardRepositoryCustom {
 }
